+++ modules/libinjection.c
@@ -83,14 +83,12 @@ ib_status_t sqli_normalize_tfn(ib_engine_t *ib,
                                ib_mpool_t *mp,
                                void *tfn_data,
                                const ib_field_t *field_in,
-                               const ib_field_t **field_out,
-                               ib_flags_t *pflags)
+                               const ib_field_t **field_out)
 {
     assert(ib != NULL);
     assert(mp != NULL);
     assert(field_in != NULL);
     assert(field_out != NULL);
-    assert(pflags != NULL);
 
     sfilter sf;
     ib_bytestr_t *bs_in;
@@ -195,8 +193,6 @@ ib_status_t sqli_normalize_tfn(ib_engine_t *ib,
         prev_token_type = current.type;
     }
 
-    /* Mark as modified. */
-    *pflags = IB_TFN_FMODIFIED;
 
     /* Create the output field wrapping bs_out. */
     buf_out_len += lead_len;