+++ CHANGES.md
@@ -52,6 +52,8 @@ IronBee v0.8.0
   Custom data fields can be indexed via the `InitVarIndexed` and
   `InitCollectionIndexed` directives.
 
+* Transformations have been simplified: Output flags have been removed.
+
 **Util**
 
 * Add external iterator support for hash.  See `ib_hash_iterator*`.